posts can only be deleted by author now

+++ src/components/Post.js
@@ -7,17 +7,13 @@ import CardActions from "@material-ui/core/CardActions";
 import ShareIcon from "@material-ui/icons/Share";
 import Button from "@material-ui/core/Button";
 import FavoriteIcon from "@material-ui/icons/Favorite";
-import Dialog from "@material-ui/core/Dialog";
-import DiaglogTitle from "@material-ui/core/DialogTitle";
-import MoreVertIcon from "@material-ui/icons/MoreVert";
 import Avatar from "@material-ui/core/Avatar";
 import IconButton from "@material-ui/core/IconButton";
+import { useSession } from "../firebase/UserProvider";
 import ReusableDialog from "../util/MyDialog";
 import Card from "@material-ui/core/Card";
 import CardHeader from "@material-ui/core/CardHeader";
-import SimpleDialogDemo from "../util/MyDialog";
-import { deletePost } from "../firebase/posts";
-import { useDocument } from "react-firebase-hooks/firestore";
+import { useCollectionData } from "react-firebase-hooks/firestore";
 import { makeStyles } from "@material-ui/core/styles";
 
 const useStyles = makeStyles((theme) => ({
@@ -36,19 +32,26 @@ const useStyles = makeStyles((theme) => ({
 }));
 
 const Post = (props) => {
+  const { user } = useSession();
   const classes = useStyles();
   const postsRef = db.collection("posts");
-  const { text, displayName, postedAt, imageURL, id } = props.post;
+  const { text, displayName, postedAt, imageURL, id, uid } = props.post;
 
   const deletePostClick = () => {
-    postsRef
-      .doc(id)
-      .delete()
-      .then(() => {
-        console.log(`document ${id} deleted successfully`);
-      });
+    if (user.uid === uid) {
+      postsRef
+        .doc(id)
+        .delete()
+        .then(() => {
+          console.log(`document ${id} deleted successfully`);
+        })
+        .catch((err) => {
+          console.log(err);
+        });
+    } else {
+      return console.log("unauthorized deletion");
+    }
   };
-
   return (
     <>
       <Card className={classes.root}>
